Refactor(RacingCar): refactor mvc pattern
